chore(assertions): split assertions into separate files (#2089)
